Renames DemoComponents screen to DemoShowroom.
