+++ boilerplate/app/navigators/DemoNavigator.tsx
@@ -5,14 +5,14 @@ import { TextStyle, ViewStyle } from "react-native"
 import { useSafeAreaInsets } from "react-native-safe-area-context"
 import { Icon } from "../components"
 import { translate } from "../i18n"
-import { DemoCommunityScreen, DemoComponentsScreen, DemoDebugScreen } from "../screens"
+import { DemoCommunityScreen, DemoShowroomScreen, DemoDebugScreen } from "../screens"
 import { DemoPodcastListScreen } from "../screens/DemoPodcastListScreen"
 import { colors, spacing, typography } from "../theme"
 import { AppStackParamList, AppStackScreenProps } from "./AppNavigator"
 
 export type DemoTabParamList = {
   DemoCommunity: undefined
-  DemoComponents: undefined
+  DemoShowroom: undefined
   DemoDebug: undefined
   DemoPodcastList: undefined
 }
@@ -45,8 +45,8 @@ export function DemoNavigator() {
       }}
     >
       <Tab.Screen
-        name="DemoComponents"
-        component={DemoComponentsScreen}
+        name="DemoShowroom"
+        component={DemoShowroomScreen}
         options={{
           tabBarLabel: translate("demoNavigator.componentsTab"),
           tabBarIcon: ({ focused }) => <Icon icon="components" color={focused && colors.tint} />,
