+++ boilerplate/app/screens/DemoShowroomScreen/demos/DemoAutoImage.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import { Image, ImageStyle, TextStyle, View, ViewStyle } from "react-native"
 import { AutoImage, Text } from "../../../components"
 import { colors, spacing } from "../../../theme"
-import { Demo } from "../DemoComponentsScreen"
+import { Demo } from "../DemoShowroomScreen"
 import { DemoDivider } from "../DemoDivider"
 import { DemoUseCase } from "../DemoUseCase"
