+++ boilerplate/app/screens/DemoShowroomScreen/demos/DemoIcon.tsx
@@ -3,7 +3,7 @@ import React from "react"
 import { ImageStyle, TextStyle, View, ViewStyle } from "react-native"
 import { Icon, iconRegistry, IconTypes, Text } from "../../../components"
 import { colors, spacing } from "../../../theme"
-import { Demo } from "../DemoComponentsScreen"
+import { Demo } from "../DemoShowroomScreen"
 import { DemoUseCase } from "../DemoUseCase"
 
 const $demoIconContainer: ViewStyle = {
