+++ boilerplate/app/screens/DemoShowroomScreen/demos/DemoListItem.tsx
@@ -4,7 +4,7 @@ import { TextStyle, View, ViewStyle } from "react-native"
 import { FlatList } from "react-native-gesture-handler"
 import { Icon, ListItem, Text } from "../../../components"
 import { colors, spacing } from "../../../theme"
-import { Demo } from "../DemoComponentsScreen"
+import { Demo } from "../DemoShowroomScreen"
 import { DemoDivider } from "../DemoDivider"
 import { DemoUseCase } from "../DemoUseCase"
