+++ boilerplate/app/screens/index.ts
@@ -3,7 +3,7 @@ export * from "./WelcomeScreen"
 export * from "./LoginScreen"
 export * from "./DemoCommunityScreen"
 export * from "./DemoDebugScreen"
-export * from "./DemoComponentsScreen/DemoComponentsScreen"
+export * from "./DemoShowroomScreen/DemoShowroomScreen"
 // @demo remove-block-end
 export * from "./ErrorScreen/ErrorBoundary"
 // export other screens here
